Fix adding cloud-init logs to diagnostics archive

+++ src/Agent.Worker/DiagnosticLogManager.cs
@@ -146,8 +146,11 @@ namespace Microsoft.VisualStudio.Services.Agent.Worker
                 string resultLogs = await DumpCloudInitLogs(logsFilePath);
                 executionContext.Debug(resultLogs);
 
-                string destination = Path.Combine(supportFilesFolder, Path.GetFileName(logsFilePath));
-                File.Copy(logsFilePath, destination);
+                if (File.Exists(logsFilePath))
+                {
+                    string destination = Path.Combine(supportFilesFolder, Path.GetFileName(logsFilePath));
+                    File.Copy(logsFilePath, destination);
+                }
 
                 executionContext.Debug("Dumping cloud-init logs is ended.");
             }
